footer style scss로 변경작업 완료
