+++ css/_footer.css
@@ -1,52 +1,52 @@
 #footer {
   height: auto;
   padding: 1rem 0 1rem 0;
-  border-top: 1px solid var(--baseColor);
-  background-color: var(--baseColor);
-}
-#footer .bind {
-  width: 50%;
-  width: 95%;
-  margin: auto;
+  border-top: 1px solid var(--bgColor);
   display: flex;
-  flex-direction: column;
   justify-content: space-between;
   align-items: center;
+  background-color: var(--bgColor);
 }
-#footer .tit {
-  width: 50%;
-  font-size: 1.5rem;
-  font-weight: 700;
-  text-align: left;
+#footer .div-left {
+  height: auto;
+  width: auto;
+  display: flex;
+  flex-direction: column;
+  margin-left: 40px;
+  margin-top: 20px;
+  background-color: var(--bgColor);
 }
-#footer h2 {
-  font-family: var(--pointFont);
-  font-size: 3rem;
-  font-weight: 500;
+#footer .div-left #footer-title {
+  color: var(--baseColor);
+  margin-left: 10px;
+}
+#footer .div-left #footer-text {
+  color: var(--baseColor);
+  margin-left: 10px;
+  margin-top: 3px;
+}
+#footer .div-right {
+  height: auto;
+  width: auto;
+  display: flex;
+  align-items: flex-end;
+  flex-direction: column;
+  margin-right: 40px;
+  margin-top: 20px;
+  background-color: var(--bgColor);
+}
+#footer .div-right .footer-clock {
   color: var(--bgColor);
 }
-#footer p {
-  padding: 0 0 10px 0;
-  display: block;
-  font-size: 1rem;
-  font-weight: 500;
-  word-break: keep-all;
+#footer .div-right .footer-weather {
   color: var(--bgColor);
+  display: flex;
+  flex-direction: column;
 }
-#footer p span {
-  font-family: var(--pointFont);
-  font-size: 2rem;
-  color: var(--bgColor);
-}
-#footer .team-parent > button {
-  min-width: 70px;
-  padding: 5px 5px;
-  margin-right: 10px;
-  margin-bottom: 10px;
-  border: 1px solid var(--bgColor);
-  border-radius: 5px;
-  font-size: 0.9rem;
-  color: var(--bgColor);
+#footer .div-right .footer-weather .icon-size {
+  width: 30px;
+  height: 30px;
+  margin-left: 10px;
 }
 
 @media (max-width: 1200px) {
