+++ css/mypage.css
@@ -171,6 +171,7 @@ html {
 }
 #header .bind .nav #search .search-input {
   width: 100%;
+  min-width: 250px;
   padding: 5px 0;
   border-bottom: 1px solid var(--baseColor);
 }
@@ -338,48 +339,63 @@ html {
   height: auto;
   padding: 1rem 0 1rem 0;
   border-top: 1px solid var(--baseColor);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
   background-color: var(--baseColor);
 }
-#footer .bind {
-  width: 95%;
-  margin: auto;
+#footer .div-left {
+  height: auto;
+  width: auto;
   display: flex;
   flex-direction: column;
+  margin-left: 40px;
+  background-color: var(--baseColor);
 }
-#footer .bind .tit {
-  width: 100%;
-  font-size: 1.5rem;
-  font-weight: 700;
-  text-align: left;
-}
-#footer .bind h2 {
-  font-family: var(--pointFont);
-  font-size: 3rem;
-  font-weight: 500;
+#footer .div-left #footer-title {
   color: var(--bgColor);
+  margin-left: 10px;
 }
-#footer .bind p {
-  padding: 0 0 10px 0;
-  display: block;
-  font-size: 1rem;
-  font-weight: 500;
-  word-break: keep-all;
+#footer .div-left #footer-text {
   color: var(--bgColor);
+  margin-left: 10px;
+  margin-top: 3px;
 }
-#footer .bind p span {
-  font-family: var(--pointFont);
-  font-size: 2rem;
+#footer .div-left .team-parent-div {
+  display: flex;
+  height: auto;
+  background-color: "var(--baseColor)";
+  margin-top: 10px;
+  margin-left: 10px;
+}
+#footer .div-left #team-people {
+  display: flex;
+  justify-content: center;
+  align-items: center;
   color: var(--bgColor);
-}
-#footer .bind .team-parent > button {
-  min-width: 70px;
-  padding: 5px 5px;
-  margin-right: 10px;
-  margin-bottom: 10px;
+  background-color: var(--baseColor);
   border: 1px solid var(--bgColor);
+  margin-right: 12px;
   border-radius: 5px;
-  font-size: 0.9rem;
+  width: 83px;
+  height: 30px;
+}
+#footer .div-right {
+  height: auto;
+  width: auto;
+  display: flex;
+  align-items: flex-end;
+  flex-direction: column;
+  margin-right: 40px;
+  background-color: var(--baseColor);
+}
+#footer .div-right .footer-clock {
   color: var(--bgColor);
+}
+#footer .div-right .footer-weather {
+  color: var(--bgColor);
+  display: flex;
+  flex-direction: column;
 }
 
 @media (max-width: 1200px) {
