+++ js/script.js
@@ -3,13 +3,13 @@ import { createSection } from './main.js';
 import { creatSlideBox } from './mainSlide.js';
 import { creatSortBox } from './sort/sort.js';
 import { createFooter } from './footer.js';
-import { createTeamIntroduction } from './teamIntroduction.js';
+// import { createTeamIntroduction } from './teamIntroduction.js';
 
 createHeader();
 creatSlideBox().then(() => {
   createSection().then(() => {
     creatSortBox();
     createFooter();
-    createTeamIntroduction();
+    // createTeamIntroduction();
   });
 });
