--- scss/_teamIntroduction.scss
@@ -1,1 +1,0 @@
-// 삭제예정, 시간과 날씨가 푸터로 옮겨가면서 파일명을 바꾸었습니다. 해당 css는  teamIntroduction.js에서 작업하였습니다.
