--- scss/review.scss
@@ -1,135 +1,0 @@
-@import 'reset';
-@import 'variables.scss';
-
-.review {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  // background-color: gray;
-  padding: 50px;
-
-  .w-60 {
-    width: 60%;
-  }
-
-  input {
-    border: 2px solid black;
-  }
-
-  h3 {
-    margin-bottom: 20px;
-    font-size: 1.8rem;
-  }
-
-  .review-content {
-    height: 100px;
-    width: 100%;
-  }
-
-  // 리뷰 중복제거
-  input {
-    margin: 10px 0;
-    padding: 10px;
-    border-radius: 5px;
-  }
-
-  button {
-    // background-color: gray;
-    margin: 10px 0;
-    padding: 10px;
-    border-radius: 5px;
-    border: 2px solid black;
-  }
-}
-
-// 리뷰 작성 form
-.review-form {
-  // background-color: green;
-
-  .form-info {
-    width: 100%;
-    display: flex;
-    justify-content: right;
-
-    .m-l-10 {
-      &:nth-child(3) {
-        background-color: var(--baseColor);
-        color: var(--bgColor);
-
-        &:hover {
-          background-color: var(--bgColor);
-          color: var(--baseColor);
-        }
-      }
-    }
-  }
-}
-
-// 리뷰 목록
-.review-list {
-  // background-color: wheat;
-}
-
-.review-item {
-  display: flex;
-  flex-direction: column;
-  border: 2px solid black;
-  margin-top: 20px;
-  padding: 0 20px;
-  border-radius: 10px;
-  background-color: white;
-
-  span {
-    margin: 10px 0;
-  }
-
-  span:nth-child(1) {
-    font-weight: bold;
-  }
-
-  span:nth-child(2) {
-    padding: 10px 0;
-  }
-}
-
-.review-form .m-l-10 {
-  margin-left: 10px;
-}
-
-// tablet
-@media (max-width: 1200px) {
-  .review {
-    padding: 20px;
-    .w-60 {
-      width: 100%;
-    }
-  }
-}
-
-// mobile
-@media (max-width: 600px) {
-  .review-form {
-    .form-info {
-      width: 100%;
-      display: flex;
-      justify-content: right;
-
-      .m-l-10 {
-        width: 38%;
-
-        &:nth-child(1) {
-          margin-left: 0;
-        }
-        &:nth-child(3) {
-          width: 20%;
-        }
-      }
-    }
-  }
-}
-
-@media (max-width: 400px) {
-  #section {
-  }
-}
